Fixed file structure issue, files should be in correct folders.
