--- hyperparameters.yml
@@ -1,40 +1,0 @@
-cartpole1:
-  env_id: CartPole-v1
-  replay_memory_size: 100000
-  mini_batch_size: 32
-  epsilon_init: 1
-  epsilon_decay: 0.9995
-  epsilon_min: 0.05
-  network_sync_rate: 10
-  learning_rate_a: 0.001
-  discount_factor_g: 0.99
-  stop_on_reward: 100000
-  fc1_nodes: 10
-
-flappybird1:
-  env_id: FlappyBird-v0
-  replay_memory_size: 100000
-  mini_batch_size: 32
-  epsilon_init: 1
-  epsilon_decay: 0.99_99_5
-  epsilon_min: 0.05
-  network_sync_rate: 10
-  learning_rate_a: 0.0001
-  discount_factor_g: 0.99
-  stop_on_reward: 100000
-  fc1_nodes: 512
-  env_make_params:
-    use_lidar: False
-
-mountaincar1:
-  env_id: MountainCar-v0
-  replay_memory_size: 100000
-  mini_batch_size: 32
-  epsilon_init: 1
-  epsilon_decay: 0.9995
-  epsilon_min: 0.05
-  network_sync_rate: 10
-  learning_rate_a: 0.001
-  discount_factor_g: 0.99
-  stop_on_reward: 100000
-  fc1_nodes: 10
