+++ rainbow_dqn_pytorch/.gitignore
@@ -1,3 +1,2 @@
+**/runs/*
 
-/runs/*
-/runs
